pyproject.toml
[project]
name = "ai-portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow>=3.0.1",
    "black[jupyter]>=25.1.0",
    "catboost>=1.2.8",
    "confluent-kafka-python>=0.0.2",
    "duckdb>=1.3.0",
    "imbalanced-learn>=0.13.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.4",
    "langgraph>=0.4.7",
    "lightgbm>=4.6.0",
    "mlflow>=2.22.0",
    "nltk>=3.9.1",
    "polars>=1.30.0",
    "pulumi>=3.173.0",
    "python-dotenv>=1.1.0",
    "quixstreams>=3.15.0",
    "scikit-learn>=1.6.1",
    "spacy>=3.8.7",
    "streamlit>=1.45.1",
    "tensorflow>=2.19.0",
    "torch>=2.7.0",
    "transformers>=4.52.3",
    "xgboost>=3.0.2",
]
